Use type name instead of having dedicated 'type' and 'state' keys.
Specifying a 'type' became mandatory with the removal of the 'typespace'
key and the value of the 'type' attribute must be a valid type name.
This vouches for yet another simplification. Since the model currently
have no keys that are valid type names (all recognized keys use lower
case letters), it is easy to identify a type name when used as a key.

Instead of requirering that a resource has both a 'type' and a 'state'
entry, this commit instead requires that the resource has exactly one
key that is a valid type name. With this change, the syntax:

    type: Aws::Vpc
    state:
      <state hash>

now becomes:

    Aws::Vpc:
      <state hash>