Allow parameter declaration without type in YAML
This commit makes it possible to declare a parameter in YAML using the
hash form without including the 'type'. This is particularly useful when
declaring a 'value' since the type can be inferred from the value type.
When neither 'type' nor 'value' is declared, or when 'value' is declared
to be null, the type will default to Any.